+++ projects/_03_independent_cqc/_03_impute/tests/fargate/utils/test_primary_service_rate_of_change.py
@@ -4,7 +4,7 @@ import polars as pl
 import polars.testing as pl_testing
 import pytest
 
-import projects._03_independent_cqc.utils.primary_service_rate_of_change.primary_service_rate_of_change as job
+import projects._03_independent_cqc._03_impute.fargate.utils.primary_service_rate_of_change as job
 from projects._03_independent_cqc.unittest_data.polars_ind_cqc_test_file_data import (
     ModelRateOfChangeData as Data,
 )
